[Refactor] pointer decls
